renamed application variables file to app-types.js & updated refs
